requirements.txt
Flask
mysql-connector
pillow
requests
evchargingfinderlib